UPDATE - renamed project to projects to be consistent with preferences
