+++ apps/authoring/src/main/models/index.ts
@@ -1,4 +1,4 @@
 export * from './model.types';
 export * from './model-service';
-export * as Project from './project';
+export * as Project from './projects';
 export * as Preferences from './preferences';
