--- apps/authoring/src/main/models/project/index.ts
@@ -1,3 +1,0 @@
-export * from './model-project.types';
-export * from './model-project';
-export * as Table from './model-projects-schema';
