+++ apps/authoring/src/main/models/projects/index.ts
@@ -1,0 +1,3 @@
+export * from './model-projects.types';
+export * from './model-projects';
+export * as Table from './model-projects-schema';
