+++ apps/authoring/src/main/models/projects/model-projects-schema.ts
