+++ apps/authoring/src/renderer/models/project/model-project.types.ts
@@ -1,4 +1,4 @@
-import { ProjectData } from '../../../main/models/project';
+import { ProjectData } from '../../../main/models/projects';
 
 export type ProjectObserverDataFn = React.Dispatch<ProjectData>;
